// corexverse/src/generated/models/Cx_ocrmenusectionsesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export const Cx_ocrmenusectionsescx_isdefault = {
  0: '___',
  1: '__'
} as const;
export type Cx_ocrmenusectionsescx_isdefault = keyof typeof Cx_ocrmenusectionsescx_isdefault;
export const Cx_ocrmenusectionsesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cx_ocrmenusectionsesstatecode = keyof typeof Cx_ocrmenusectionsesstatecode;
export const Cx_ocrmenusectionsesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cx_ocrmenusectionsesstatuscode = keyof typeof Cx_ocrmenusectionsesstatuscode;

export interface Cx_ocrmenusectionses {
  createdby?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfby?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  cx_color?: string;
  cx_description?: string;
  cx_displayorder?: string;
  cx_isdefault?: Cx_ocrmenusectionsescx_isdefault;
  cx_isdefaultname?: string;
  cx_name?: string;
  cx_ocrmenusectionid: string;
  cx_ocrmenusectionsid: string;
  importsequencenumber?: string;
  modifiedby?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfby?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunit?: string;
  owningbusinessunitname: string;
  owningteam?: string;
  owninguser?: string;
  statecode: Cx_ocrmenusectionsesstatecode;
  statecodename?: string;
  statuscode?: Cx_ocrmenusectionsesstatuscode;
  statuscodename?: string;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
  versionnumber?: string;
}