// corexverse/src/generated/models/Cx_ocrfoldersModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export const Cx_ocrfoldersstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cx_ocrfoldersstatecode = keyof typeof Cx_ocrfoldersstatecode;
export const Cx_ocrfoldersstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cx_ocrfoldersstatuscode = keyof typeof Cx_ocrfoldersstatuscode;

export interface Cx_ocrfolders {
  createdby?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfby?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  cx_description?: string;
  cx_documentcount?: string;
  cx_foldercount?: string;
  cx_menusectionid?: string;
  cx_menusectionidname?: string;
  cx_name?: string;
  cx_ocrfolderid: string;
  cx_ocrfolderid1: string;
  cx_parentfolderid?: string;
  cx_parentfolderidname?: string;
  cx_path?: string;
  importsequencenumber?: string;
  modifiedby?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfby?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunit?: string;
  owningbusinessunitname: string;
  owningteam?: string;
  owninguser?: string;
  statecode: Cx_ocrfoldersstatecode;
  statecodename?: string;
  statuscode?: Cx_ocrfoldersstatuscode;
  statuscodename?: string;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
  versionnumber?: string;
}