// corexverse/src/generated/models/Cx_ocrdocumentsesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export const Cx_ocrdocumentsescx_status = {
  0: '________',
  1: '____',
  2: '___',
  3: '__',
  4: '___'
} as const;
export type Cx_ocrdocumentsescx_status = keyof typeof Cx_ocrdocumentsescx_status;
export const Cx_ocrdocumentsesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cx_ocrdocumentsesstatecode = keyof typeof Cx_ocrdocumentsesstatecode;
export const Cx_ocrdocumentsesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cx_ocrdocumentsesstatuscode = keyof typeof Cx_ocrdocumentsesstatuscode;

export interface Cx_ocrdocumentses {
  createdby?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfby?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  cx_description?: string;
  cx_filedata?: string;
  cx_filedata_name?: string;
  cx_filename: string;
  cx_filesize: string;
  cx_filetype: string;
  cx_fileurl: string;
  cx_folderid?: string;
  cx_folderidname?: string;
  cx_name: string;
  cx_ocrdocumentsid: string;
  cx_status?: Cx_ocrdocumentsescx_status;
  cx_statusname?: string;
  cx_tags?: string;
  cx_thumbnailurl?: string;
  cx_uploadedby?: string;
  cx_uploadedbyname?: string;
  cx_uploadedbyyominame?: string;
  cx_uploadeddate?: string;
  importsequencenumber?: string;
  modifiedby?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfby?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunit?: string;
  owningbusinessunitname: string;
  owningteam?: string;
  owninguser?: string;
  statecode: Cx_ocrdocumentsesstatecode;
  statecodename?: string;
  statuscode?: Cx_ocrdocumentsesstatuscode;
  statuscodename?: string;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
  versionnumber?: string;
}